RavenDB-17973 QueryFilter moved into a separate file.
