+++ src/Raven.Server/Documents/Indexes/Persistence/Lucene/IndexOperationBase.cs
@@ -7,7 +7,6 @@ using System.Runtime.CompilerServices;
 using Lucene.Net.Analysis;
 using Lucene.Net.Search;
 using Lucene.Net.Store;
-using Nest;
 using Raven.Client;
 using Raven.Client.Documents.Indexes;
 using Raven.Server.Documents.Indexes.Persistence.Lucene.Analyzers;
@@ -280,72 +279,7 @@ namespace Raven.Server.Documents.Indexes.Persistence.Lucene
             if (query.Metadata.FilterScript is null)
                 return null;
 
-            return new QueryFilter(_index, query, documentsContext, skippedResults, scannedDocuments, retriever, queryTimings);
-        }
-        
-        protected enum FilterResult
-        {
-            Accepted,
-            Skipped,
-            LimitReached
-        }
-        protected class QueryFilter : IDisposable
-        {
-            private readonly IndexQueryServerSide _query;
-            private readonly DocumentsOperationContext _documentsContext;
-            private readonly Reference<int> _skippedResults;
-            private readonly Reference<int> _scannedDocuments;
-            private readonly IQueryResultRetriever _retriever;
-            private readonly QueryTimingsScope _queryTimings;
-            private readonly ScriptRunner.SingleRun _filterScriptRun;
-            private ScriptRunner.ReturnRun _filterSingleRun;
-            
-            public QueryFilter(Index index, IndexQueryServerSide query, DocumentsOperationContext documentsContext, Reference<int> skippedResults, Reference<int> scannedDocuments, IQueryResultRetriever retriever,QueryTimingsScope queryTimings)
-            {
-                _query = query;
-                _documentsContext = documentsContext;
-                _skippedResults = skippedResults;
-                _scannedDocuments = scannedDocuments;
-                _retriever = retriever;
-                _queryTimings = queryTimings;
-                
-                var key = new CollectionQueryEnumerable.FilterKey(query.Metadata);
-                _filterSingleRun = index.DocumentDatabase.Scripts.GetScriptRunner(key, readOnly: true, patchRun: out _filterScriptRun);
-            }
-
-            public FilterResult Apply(global::Lucene.Net.Documents.Document document, string key, IState state)
-            {
-                var doc = _retriever.DirectGet(document, key, DocumentFields.All, state);
-                if (doc == null)
-                {
-                    _skippedResults.Value++;
-                    return FilterResult.Skipped;
-                }
-
-                if (_scannedDocuments.Value >= _query.FilterLimit)
-                {
-                    return FilterResult.LimitReached;
-                }
-                
-                _scannedDocuments.Value++;
-                object self = _filterScriptRun.Translate(_documentsContext, doc);
-                using(_queryTimings?.For(nameof(QueryTimingsScope.Names.Filter)))
-                using (var result = _filterScriptRun.Run(_documentsContext, _documentsContext, "execute", new[]{self, _query.QueryParameters}, _queryTimings))
-                {
-                    if (result.BooleanValue == true)
-                    {
-                        return FilterResult.Accepted;
-                    }
-                
-                    _skippedResults.Value++;
-                    return FilterResult.Skipped;
-                }
-            }
-            
-            public void Dispose()
-            {
-                _filterSingleRun.Dispose();
-            }
+            return new QueryFilter(index, query, documentsContext, skippedResults, scannedDocuments, retriever, queryTimings);
         }
     }
 }
